+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/model/Channel.kt
@@ -8,22 +8,22 @@ import kotlinx.serialization.internal.StringDescriptor
 data class Channel(
     @SerialName("id") val id: String,
     @SerialName("type") val type: ChannelType,
-    @Optional @SerialName("guild_id") val guildId: String? = null,
-    @Optional @SerialName("position") var position: Int? = null,
-    @Optional @SerialName("permission_overwrites") var permissionOverwrites: List<Overwrite> = emptyList(),
-    @Optional @SerialName("name") var name: String? = null,
-    @Optional @SerialName("topic") var topic: String? = null,
-    @Optional @SerialName("nsfw") var nsfw: Boolean? = null,
-    @Optional @SerialName("last_message_id") var lastMessageId: String? = null,
-    @Optional @SerialName("bitrate") var voiceBitrate: Int? = null,
-    @Optional @SerialName("user_limit") var voiceUserLimit: Int? = null,
-    @Optional @SerialName("rate_limit_per_user") var rateLimitPerUser: Int? = null,
-    @Optional @SerialName("recipients") var dmRecipients: List<User>? = null,
-    @Optional @SerialName("icon") var iconHash: String? = null,
-    @Optional @SerialName("owner_id") var dmOwnerId: String? = null,
-    @Optional @SerialName("application_id") var applicationId: String? = null,
-    @Optional @SerialName("parent_id") var parentId: String? = null,
-    @Optional @SerialName("last_pin_timestamp") var lastPinTime: String? = null
+    @SerialName("guild_id") val guildId: String? = null,
+    @SerialName("position") var position: Int? = null,
+    @SerialName("permission_overwrites") var permissionOverwrites: List<Overwrite> = emptyList(),
+    @SerialName("name") var name: String? = null,
+    @SerialName("topic") var topic: String? = null,
+    @SerialName("nsfw") var nsfw: Boolean? = null,
+    @SerialName("last_message_id") var lastMessageId: String? = null,
+    @SerialName("bitrate") var voiceBitrate: Int? = null,
+    @SerialName("user_limit") var voiceUserLimit: Int? = null,
+    @SerialName("rate_limit_per_user") var rateLimitPerUser: Int? = null,
+    @SerialName("recipients") var dmRecipients: List<User>? = null,
+    @SerialName("icon") var iconHash: String? = null,
+    @SerialName("owner_id") var dmOwnerId: String? = null,
+    @SerialName("application_id") var applicationId: String? = null,
+    @SerialName("parent_id") var parentId: String? = null,
+    @SerialName("last_pin_timestamp") var lastPinTime: String? = null
 )
 
 @Serializable(with = ChannelTypeSerializer::class)