+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/websocket/events/GuildMemberAdd.kt
@@ -1,7 +1,6 @@
 package com.jessecorbett.diskord.api.websocket.events
 
 import com.jessecorbett.diskord.api.model.User
-import kotlinx.serialization.Optional
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -10,8 +9,8 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class GuildMemberAdd(
-    @Optional @SerialName("user") val user: User? = null,
-    @Optional @SerialName("nick") val nickname: String? = null,
+    @SerialName("user") val user: User? = null,
+    @SerialName("nick") val nickname: String? = null,
     @SerialName("roles") val roleIds: List<String>,
     @SerialName("joined_at") val joinedAt: String,
     @SerialName("deaf") val isDeaf: Boolean,