+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/websocket/events/MessageReactionRemoveAll.kt
@@ -1,6 +1,5 @@
 package com.jessecorbett.diskord.api.websocket.events
 
-import kotlinx.serialization.Optional
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -8,5 +7,5 @@ import kotlinx.serialization.Serializable
 data class MessageReactionRemoveAll(
     @SerialName("channel_id") val channelId: String,
     @SerialName("message_id") val messageId: String,
-    @Optional @SerialName("guild_id") val guildId: String? = null
+    @SerialName("guild_id") val guildId: String? = null
 )