+++ src/commonMain/kotlin/com/jessecorbett/diskord/internal/CommonResources.kt
@@ -5,7 +5,7 @@ import io.ktor.client.engine.HttpClientEngineConfig
 import io.ktor.client.engine.HttpClientEngineFactory
 
 internal const val defaultUserAgentUrl = "https://gitlab.com/jesselcorbett/diskord"
-internal const val defaultUserAgentVersion = "1.4.0"
+internal const val defaultUserAgentVersion = "1.4.1"
 
 internal expect fun websocketClient(): HttpClientEngineFactory<HttpClientEngineConfig>
 internal expect fun httpClient(): HttpClientEngineFactory<HttpClientEngineConfig>